Update runner state module and documentation
